requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.6
django-dotenv==1.4.2
djangorestframework==3.12.4
idna==2.10
numpy==1.21.2
pandas==1.1.5
pip==21.2.2
psycopg2-binary
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.25.1
setuptools==57.4.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
wheel==0.36.2
xlrd==1.2.0